refactor: global 아키텍처 수정
