+++ src/main/java/org/sopt/global/common/response/ErrorResponse.java
@@ -1,3 +1,3 @@
-package org.sopt.global.response;
+package org.sopt.global.common.response;
 
 public record ErrorResponse(String message) {}
