+++ src/main/java/org/sopt/global/exception/BoardTypeNotFoundException.java
@@ -1,6 +1,6 @@
 package org.sopt.global.exception;
 
-import org.sopt.global.code.ErrorCode;
+import org.sopt.global.common.code.ErrorCode;
 
 // 존재하지 않는 board type 을 찾을 때
 public class BoardTypeNotFoundException extends BaseException {
