+++ src/main/java/org/sopt/like/controller/LikeController.java
@@ -3,8 +3,8 @@ package org.sopt.like.controller;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import org.sopt.like.dto.request.CreateLikePostRequest;
-import org.sopt.global.code.SuccessCode;
-import org.sopt.global.response.BaseResponse;
+import org.sopt.global.common.code.SuccessCode;
+import org.sopt.global.common.response.BaseResponse;
 import org.sopt.like.service.LikeService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
