+++ src/main/java/org/sopt/post/controller/PostController.java
@@ -10,8 +10,8 @@ import org.sopt.post.dto.request.CreatePostRequest;
 import org.sopt.post.dto.request.UpdatePostRequest;
 import org.sopt.post.dto.response.CreatePostResponse;
 import org.sopt.post.dto.response.PostResponse;
-import org.sopt.global.code.SuccessCode;
-import org.sopt.global.response.BaseResponse;
+import org.sopt.global.common.code.SuccessCode;
+import org.sopt.global.common.response.BaseResponse;
 import org.sopt.post.service.PostService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
