+++ src/main/java/org/sopt/user/controller/UserController.java
@@ -3,8 +3,8 @@ package org.sopt.user.controller;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import org.sopt.user.dto.request.CreateUserRequest;
-import org.sopt.global.code.SuccessCode;
-import org.sopt.global.response.BaseResponse;
+import org.sopt.global.common.code.SuccessCode;
+import org.sopt.global.common.response.BaseResponse;
 import org.sopt.user.service.UserService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
